requirements.txt
pinax
pinax-ratings
Django
django-star-ratings
django-crispy-forms
django-markdown-deux
django-pagedown
markdown2
olefile
Pillow
pytz
django-import-export